requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
filelock==3.18.0
fsspec==2025.3.0
fuzzywuzzy==0.18.0
gensim==0.10.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
Levenshtein==0.27.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
pillow==11.1.0
python-Levenshtein==0.27.1
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
